requirements.txt
streamlit
requests
groq
deepgram-sdk
chromadb
langchain
sentence-transformers
langchain-community
pysqlite3-binary